config: add optional CERT_BUNDLE argument (#232)
This is useful when needing to pass a custom Certificate Bundle. For
instance Proton Mail requires you to run a local software called
Proton Mail Bridge to act as a proxy between your machine and Proton
Mail servers. To correctly authenticate, you need export a TLS
certificate from Proton Mail Bridge and pass it to the Config object
in FastAPI Mail, so that aiosmtplib can work correctly.